move logger out of __init__
