+++ skellycam/__init__.py
@@ -9,11 +9,7 @@ __package_name__ = "skellycam"
 __repo_url__ = f"https://github.com/freemocap/{__package_name__}/"
 __repo_issues_url__ = f"{__repo_url__}issues"
 
-import logging
-
 from skellycam.system.environment.configure_logging import configure_logging, LogLevel
 
 configure_logging(LogLevel.TRACE)
-
-logger = logging.getLogger(__name__)
