+++ skellycam/backend/backend_main.py
@@ -1,7 +1,7 @@
 import multiprocessing
 from multiprocessing import Event
 
-from skellycam import logger
+from skellycam.system.environment.get_logger import logger
 from skellycam.backend.backend_loop import backend_loop
 from skellycam.utilities.clean_up import remove_empty_directories
