+++ skellycam/backend/controller/core_functionality/camera_group/strategies/camera_subarray_pipe_process.py
@@ -5,7 +5,7 @@ from typing import Dict, List, Any, Optional
 
 from setproctitle import setproctitle
 
-from skellycam import logger
+from skellycam.system.environment.get_logger import logger
 from skellycam.backend.controller.core_functionality.camera_group.camera.camera import Camera
 from skellycam.models.cameras.camera_config import CameraConfig
 from skellycam.models.cameras.camera_id import CameraId
