+++ skellycam/backend/controller/core_functionality/camera_group/video_recorder/timestamps/timestamp_logger_manager.py
@@ -5,7 +5,7 @@ from typing import Dict, Optional, Tuple, List, Any, Hashable
 
 import pandas as pd
 
-from skellycam import logger
+from skellycam.system.environment.get_logger import logger
 from skellycam.backend.controller.core_functionality.camera_group.video_recorder.timestamps.timestamp_logger import \
     CameraTimestampLogger
 from skellycam.models.cameras.camera_config import CameraConfig
