+++ skellycam/frontend/application/qt_application.py
@@ -1,6 +1,6 @@
 from PySide6.QtWidgets import QApplication
 
-from skellycam import logger
+from skellycam.system.environment.get_logger import logger
 
 _QT_APPLICATION = None
