+++ skellycam/frontend/gui/main_window/keyboard_shortcuts.py
@@ -2,7 +2,7 @@ import multiprocessing
 
 from PySide6.QtGui import QShortcut, QKeySequence
 
-from skellycam import logger
+from skellycam.system.environment.get_logger import logger
 
 
 class KeyboardShortcuts:
