+++ skellycam/frontend/gui/skellycam_widget/helpers/backend_communicator.py
@@ -3,7 +3,7 @@ from typing import Callable
 
 from PySide6.QtCore import QTimer
 
-from skellycam import logger
+from skellycam.system.environment.get_logger import logger
 from skellycam.backend.controller.interactions.base_models import BaseResponse, BaseInteraction
