+++ skellycam/system/environment/get_logger.py
@@ -1,0 +1,3 @@
+import logging
+
+logger = logging.getLogger(__name__)
