+++ skellycam/utilities/clean_up.py
@@ -1,5 +1,4 @@
-
-from skellycam import logger
+from skellycam.system.environment.get_logger import logger
 from skellycam.system.environment.default_paths import get_default_skellycam_base_folder_path
